Python-Matplotlib
Matplotlib is a popular Python library used for creating static, interactive, and animated visualizations. It provides a wide variety of plotting options and customization capabilities, making it a powerful tool for data visualization. Matplotlib is often used in data analysis, scientific research, and various other fields to represent data in graphical form.