Scikit-learn python
Scikit-learn (or sklearn) is a popular open-source machine learning library for Python. It provides a wide range of tools and algorithms for various machine learning tasks, including classification, regression, clustering, dimensionality reduction, and more. Scikit-learn is built on top of other scientific Python libraries like NumPy, SciPy, and matplotlib, making it a powerful and easy-to-use tool for machine learning projects.